requirements.txt
fastapi
openai
pydantic
modal
streamlit
python-dotenv